requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
psycopg2-binary
PyJWT
requests
flask-cors
